experimental drive code

+++ src/utils.cpp
@@ -24,7 +24,7 @@ int inchesToCounts(float inches)
     return static_cast<int>(inches * (COUNTS_PER_ENCODER / (M_PI * WHEEL_DIAMETER)));
 }
 
-float countsToInches(int counts)
+float countsToInches(float counts)
 {
     return counts / (COUNTS_PER_ENCODER / (M_PI * WHEEL_DIAMETER));
 }
